+++ openccjava/src/main/java/openccjava/ConversionPlanCache.java
@@ -132,8 +132,9 @@ public final class ConversionPlanCache {
      *   <li><b>S2T / T2S</b> – Simplified ↔ Traditional with optional punctuation dictionaries</li>
      *   <li><b>S2Tw / Tw2S / S2Twp / Tw2Sp</b> – Conversions involving Taiwan-specific
      *       phrase and variant dictionaries</li>
-     *   <li><b>S2Hk / S2Hkp / Hk2S / Hk2Sp / T2Hk / Hk2T</b> – Conversions involving Hong Kong variants</li>
+     *   <li><b>S2Hk / S2Hkp / Hk2S / Hk2Sp</b> – Simplified ↔ Hong Kong conversions</li>
      *   <li><b>T2Tw / T2Twp / Tw2T / Tw2Tp</b> – Traditional ↔ Taiwan conversions</li>
+     *   <li><b>T2Hk / T2Hkp / Hk2T / Hk2Tp</b> – Traditional ↔ Hong Kong conversions</li>
      *   <li><b>T2Jp / Jp2T</b> – Traditional ↔ Japanese conversions</li>
      * </ul>
      *
@@ -261,9 +262,19 @@ public final class ConversionPlanCache {
                 refs = new DictRefs(r1, unionCache.unionFor(UnionKey.HkVariantsPair));
                 break;
             }
+            case T2HKP: {
+                r1 = Arrays.asList(d.hk_phrases, d.hk_variants_phrases, d.hk_variants);
+                refs = new DictRefs(r1, unionCache.unionFor(UnionKey.HkTriple));
+                break;
+            }
             case HK2T: {
                 r1 = Arrays.asList(d.hk_variants_rev_phrases, d.hk_variants_rev);
                 refs = new DictRefs(r1, unionCache.unionFor(UnionKey.HkRevPair));
+                break;
+            }
+            case HK2TP: {
+                r1 = Arrays.asList(d.hk_phrases_rev, d.hk_variants_rev_phrases, d.hk_variants_rev);
+                refs = new DictRefs(r1, unionCache.unionFor(UnionKey.HkRevTriple));
                 break;
             }
             case T2JP: {